Moving out of Linq folder
