+++ source/MasterDevs.Core.Tests/MasterDevs.Core.Tests.csproj
@@ -54,8 +54,8 @@
     <Compile Include="System\ActionTests.cs" />
     <Compile Include="System\ClassExtensionsTests.cs" />
     <Compile Include="System\EventHandlerExtensionsTests.cs" />
-    <Compile Include="System\Linq\IEnumerableTests.cs" />
-    <Compile Include="System\Linq\ObjectTests.cs" />
+    <Compile Include="System\IEnumerableTests.cs" />
+    <Compile Include="System\ObjectTests.cs" />
     <Compile Include="System\NullableExtensionsTests.cs" />
     <Compile Include="System\StringExtensionsTests.cs" />
   </ItemGroup>
