+++ source/MasterDevs.Core/System/ObjectExtensions.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 
-namespace System.Linq
+namespace System
 {
     public static class ObjectExtensions
     {
@@ -13,6 +13,16 @@ namespace System.Linq
         public static bool In<TSource>(this TSource value, IEnumerable<TSource> source, IEqualityComparer<TSource> comparer)
         {
             return source.Contains(value, comparer);
+        }
+
+        public static string ToStringSafe(this object value, string defaultValue = @"")
+        {
+            if (null == value)
+            {
+                return null == defaultValue ? string.Empty : defaultValue;
+            }
+
+            return value.ToString();
         }
     }
 }
